main_server/classifier.py
import torch
import torchvision.transforms as transforms
from torchvision import models
from PIL import Image
import os

# Basic config
MODEL_PATH = "model.pth"  # fine-tuned ResNet model


def load_model():
    model = models.resnet18(pretrained=False)
    model.load_state_dict(torch.load(MODEL_PATH, map_location=torch.device("cpu")))
    model.eval()
    return model


model = load_model()

transform = transforms.Compose(
    [
        transforms.Resize((224, 224)),
        transforms.ToTensor(),
        transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225]),
    ]
)


def classify_image(image_path: str) -> int:
    if not os.path.exists(image_path):
        raise FileNotFoundError(f"Image not found: {image_path}")

    image = Image.open(image_path).convert("RGB")
    tensor = transform(image).unsqueeze(0)

    with torch.no_grad():
        outputs = model(tensor)
        _, predicted = torch.max(outputs, 1)
        return int(predicted.item())
